+++ client/src/Component/DetailPage/ReviewItem.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import Button from "../style/button";
+import Button from "../style/StyleButton";
 import ImgBtn from "../style/ImgBtn";
 import profile from "../style/img/profile.png";
