+++ client/src/Component/DetailPage/StoreInfo.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import MenuItem from "./MenuItem";
-import Button from "./../style/button";
+import Button from "../style/StyleButton";
 
 const Container = styled.div`
   display: flex;
